--- nym-vpn-core/crates/nym-vpnd/src/command_interface/error.rs
@@ -1,19 +1,0 @@
-// Copyright 2024 - Nym Technologies SA <contact@nymtech.net>
-// SPDX-License-Identifier: GPL-3.0-only
-
-use std::path::PathBuf;
-
-#[derive(Debug, thiserror::Error)]
-pub enum CommandInterfaceError {
-    #[error("failed to parse DNS IP address: {ip}")]
-    ParseDnsIp {
-        ip: String,
-        source: std::net::AddrParseError,
-    },
-
-    #[error("failed to create incoming stream at {}", socket_path.display())]
-    CreateIncoming {
-        socket_path: PathBuf,
-        source: std::io::Error,
-    },
-}
