--- nym-vpn-core/crates/nym-vpnd/src/command_interface/mod.rs
@@ -1,9 +1,0 @@
-// Copyright 2024 - Nym Technologies SA <contact@nymtech.net>
-// SPDX-License-Identifier: GPL-3.0-only
-
-mod config;
-mod error;
-mod listener;
-mod start;
-
-pub use start::start_command_interface;
